+++ sperl_hash_func.c
@@ -5,7 +5,7 @@
 #include "sperl.h"
 #include "sperl_hash_func.h"
 
-uint32_t SPerl_HASH_FUNC_calc_hash(SPerl* sperl, const char* str, size_t len) {
+uint32_t SPerl_HASH_FUNC_calc_hash(SPerl* sperl, const char* str, int32_t len) {
   (void)sperl;
   
   assert(len > 0);
